requirements.txt
pandas==2.2.3
transformers==4.46.0
openai==1.52.2
pytest==7.3.1
python-dotenv==1.0.1
bert-score==0.3.13
scikit_learn==1.0.2
matplotlib==3.9.2
torch==2.2.2
tqdm==4.66.6
requests==2.32.3
numpy==1.24.3
pydantic==2.9.2
anyio==4.6.2.post1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
safetensors==0.4.5
